Revert to .net6 version
Can not upgrade to .net9 since grapical component StockSharp is still in .net6

+++ Common/Api/ObjectStoreResponse.cs
@@ -130,7 +130,7 @@ namespace QuantConnect.Api
         /// Size of all objects stored in bytes
         /// </summary>
         [JsonProperty(PropertyName = "objectStorageUsed")]
-        public long ObjectStorageUsed { get; set; }
+        public int ObjectStorageUsed { get; set; }
 
         /// <summary>
         /// Size of all the objects stored in human-readable format
